computers/get/entry now uses Spirit::Computer to create and get computer entries. Spirit::Computer.create() creates a hash based upon computer defaults. Spirit::Computer now has an instance definition and is used to read individual computer plists. Fixed Spirit::Log not checking for the existence of a log before attempting to rotate existing logs. Added PlistModel to_hash method as syntactic sugar for retrieving plist contents as a hash. DS Runtime can now successfully create and query computer items.

+++ lib/spirit/log.rb
@@ -11,23 +11,25 @@ module Spirit
     end
 
     def rotate!
-      rotated_log = File.join(self.class.path, 'rotate', name+@ext)
+      if File.exists? @path
+        rotated_log = File.join(self.class.path, 'rotate', name+@ext)
 
-      # Rotate old log if necessary
-      if File.exists? rotated_log
-        i = 0
+        # Rotate old log if necessary
+        if File.exists? rotated_log
+          i = 0
 
-        while File.exists? File.join(self.class.path, 'rotate', @name + '.' + i + @ext)
-          i = i.next
+          while File.exists? File.join(self.class.path, 'rotate', @name + '.' + i + @ext)
+            i = i.next
+          end
+
+          next_log_name = File.join(self.class.path, 'rotate', @name + '.' + i + @ext)
+
+          File.rename(rotated_log, next_log_name)
         end
 
-        next_log_name = File.join(self.class.path, 'rotate', @name + '.' + i + @ext)
-
-        File.rename(rotated_log, next_log_name)
+        # Now rotate existing log to rotate/serial.log
+        File.rename(@path, rotated_log)
       end
-
-      # Now rotate existing log to rotate/serial.log
-      File.rename(@path, rotated_log)
     end
 
   end
